Muuta Float-datatyypit Doubleiksi
Scala-scheman käyttämä Json4s parseri asettaa floatit doubleiksi jolloin
esim. 4.6:n esitys muuttuu 4.59999991223. Käyttämällä Doubleja
kierretään tämä ongelma

+++ src/main/scala/fi/oph/koski/documentation/ExamplesDIA.scala
@@ -190,7 +190,7 @@ object ExamplesDIA {
   def diaTutkintovaiheenSuoritus(kokonaispistemäärä: Option[Int] = None,
                                  lukukausisuoritustenKokonaispistemäärä: Option[Int] = None,
                                  tutkintoaineidenKokonaispistemäärä: Option[Int] = None,
-                                 kokonaispistemäärästäJohdettuKeskiarvo: Option[Float] = None) = DIATutkinnonSuoritus(
+                                 kokonaispistemäärästäJohdettuKeskiarvo: Option[Double] = None) = DIATutkinnonSuoritus(
     toimipiste = saksalainenKoulu,
     suorituskieli = englanti,
     vahvistus = ExampleData.vahvistusPaikkakunnalla(org = saksalainenKoulu, kunta = helsinki),
