+++ src/main/scala/fi/oph/koski/raportit/LukioRaportti.scala
@@ -322,7 +322,7 @@ case class LukioOppiaineenKurssienVälilehtiStaattisetKolumnit(
   suorituksenTyyppi: String
 )
 
-case class LukioKurssinTiedot(kurssintyyppi: Option[String], arvosana: Option[String], laajuus: Option[Float], tunnustettu: Boolean) {
+case class LukioKurssinTiedot(kurssintyyppi: Option[String], arvosana: Option[String], laajuus: Option[Double], tunnustettu: Boolean) {
   override def toString: String = s"${kurssintyyppi.getOrElse("Ei tyyppiä")},${arvosana.map("Arvosana " + _).getOrElse("Ei arvosanaa")},${laajuus.map("Laajuus " + _).getOrElse("Ei laajuutta")}${if (tunnustettu) ",tunnustettu}" else ""}"
 }
